msm: qpnp-power-on: Ignore input calls from vol-down key
The power button is handled through this driver, but for some reason pressing the power button always triggers the vol-down button, thus whenever the power button is pressed, Android interprets it as if the power button and vol-down button are being pressed simultaneously. Ignore calls to the vol-down button from here to fix.

Signed-off-by: Sultanxda <sultanxda@gmail.com>